+++ core/systems/storage/_routes.ex
@@ -1,0 +1,10 @@
+defmodule Systems.Storage.Routes do
+  defmacro routes() do
+    quote do
+      scope "/", Systems.Storage do
+        pipe_through([:browser, :require_authenticated_user])
+        live("/storage/:id/content", ContentPage)
+      end
+    end
+  end
+end