+++ core/systems/storage/built_in/endpoint_model.ex
@@ -14,8 +14,7 @@ defmodule Systems.Storage.BuiltIn.EndpointModel do
   end
 
   def changeset(endpoint, params) do
-    endpoint
-    |> cast(params, @fields)
+    endpoint |> cast(params, @fields)
   end
 
   def validate(changeset) do